Add files via upload
Changes made:

At least 1 Operator: Although the code does not explicitly use arithmetic or comparison operators, it still employs the assignment operator (=) and the concatenation operator (.). For example, in the foreach loop to display tasks, the concatenation operator (.) is used to concatenate strings for HTML output.

At least 1 Data Structure: The code uses the SplStack data structure, which is a stack implemented in the Standard PHP Library (SPL). This data structure is used to store and manage the todo list tasks.

At least 1 User Defined Function: The code defines two user-defined functions: appendData() and deleteData(). These functions are responsible for adding tasks to the todo list and deleting tasks from the todo list, respectively.

At least 1 System Defined Function: The code utilizes several system-defined functions, such as session_start() to start a session, isset() to check if a variable is set, empty() to check if a variable is empty, urlencode() to encode a string for use in a URL, and various functions provided by the Standard PHP Library (SPL), such as SplStack methods to manipulate the stack data structure.